+++ src/server/src/main/java/com/drm/server/controller/client/MediaController.java
@@ -1,26 +1,72 @@
-package com.drm.server.controller;
+package com.drm.server.controller.client;
 
+import com.drm.server.common.APIResponse;
 import com.drm.server.common.ErrorResponse;
+import com.drm.server.common.enums.SuccessCode;
+import com.drm.server.controller.dto.request.MediaRequest;
+import com.drm.server.controller.dto.response.DashboardResponse;
+import com.drm.server.controller.dto.response.MediaApplicationResponse;
+import com.drm.server.controller.dto.response.MediaResponse;
+import com.drm.server.domain.dashboard.Dashboard;
+import com.drm.server.domain.location.Location;
+import com.drm.server.domain.media.Media;
+import com.drm.server.domain.mediaApplication.MediaApplication;
+import com.drm.server.domain.user.CustomUserDetails;
+import com.drm.server.domain.user.User;
+import com.drm.server.service.*;
+import com.google.protobuf.Api;
+import io.swagger.v3.oas.annotations.Hidden;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
 
 @RestController
-@RequestMapping("api/v1/dashboard")
+@RequestMapping("api/v1/media")
 @RequiredArgsConstructor
-@Tag(name = "Dashboard",description = "대시보드 관련 API")
+@Tag(name = "Media",description = "광고 등록 관련 api")
 @CrossOrigin(origins = "http://localhost:5173", allowedHeaders = "x-requested-with, Authorization, Content-Type")
-public class DashBoardController {
-    @GetMapping()
-    @Operation(summary = "전체 대시보드 리스트 조회")
+public class MediaController {
+    private final UserService userService;
+    private final DashboardService dashboardService;
+    private final MediaService mediaService;
+    private final MediaApplicationService mediaApplicationService;
+    private final LocationService locationService;
+    @Operation(summary = "광고 등록")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "성공"),
+            @ApiResponse(responseCode = "400", description = "요청 형식 혹은 요청 콘텐츠가 올바르지 않을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+            @ApiResponse(responseCode = "401", description = "토큰 시간 만료, 형식 오류,로그아웃한 유저 접근,헤더에 값이 없을때",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+            @ApiResponse(responseCode = "403", description = "권한이 없는 경우",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+            @ApiResponse(responseCode = "404", description = "요청한 URL/URI와 일치하는 항목을 찾지 못함,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+            @ApiResponse(responseCode = "500", description = "외부 API 요청 실패, 정상적 수행을 할 수 없을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+    })
+    @PostMapping(path = "",consumes =  {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
+    public ResponseEntity<APIResponse<MediaApplicationResponse.TotalApplicationInfo>> createMedia(@RequestPart(value = "request") @Valid MediaRequest.Create create, @RequestPart(value = "file",required = false) MultipartFile multipartFile, @AuthenticationPrincipal CustomUserDetails userDetails){
+        User getUser =userService.getUser(userDetails.getUsername());
+        Location location = locationService.findById(create.getLocationId());
+
+        Dashboard dashboard = dashboardService.createDashboard(create, getUser);
+        Media media=  mediaService.createMedia(create, dashboard,multipartFile);
+        MediaApplication mediaApplication = mediaApplicationService.createMediaApplication(media,location,create.getStartDate(),create.getEndDate());
+
+        MediaApplicationResponse.TotalApplicationInfo applicationInfo = new MediaApplicationResponse.TotalApplicationInfo(mediaApplication);
+        APIResponse response = APIResponse.of(SuccessCode.INSERT_SUCCESS, applicationInfo);
+        return new ResponseEntity<>(response, HttpStatus.CREATED);
+    }
+    @Operation(summary = "등록한 광고 리스트")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "성공"),
             @ApiResponse(responseCode = "400", description = "요청 형식 혹은 요청 콘텐츠가 올바르지 않을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
@@ -29,11 +75,29 @@ public class DashBoardController {
             @ApiResponse(responseCode = "404", description = "요청한 URL/URI와 일치하는 항목을 찾지 못함,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
             @ApiResponse(responseCode = "500", description = "외부 API 요청 실패, 정상적 수행을 할 수 없을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
     })
-    public void getDashboards(){
+    @GetMapping("")
+    public ResponseEntity<APIResponse<List<MediaResponse.MediaInfo>>> getMedia(@AuthenticationPrincipal CustomUserDetails userDetails){
+        User getUser = userService.getUser(userDetails.getUsername());
+        List<Dashboard> dashboards = dashboardService.findByUser(getUser);
+        List<MediaResponse.MediaInfo> mediaInfos = mediaService.findByDashboard(dashboards);
+        APIResponse response = APIResponse.of(SuccessCode.SELECT_SUCCESS, mediaInfos);
+        return new ResponseEntity<>(response, HttpStatus.OK);
+    }
+    @Operation(summary = "광고 삭제")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "400", description = "요청 형식 혹은 요청 콘텐츠가 올바르지 않을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+            @ApiResponse(responseCode = "401", description = "토큰 시간 만료, 형식 오류,로그아웃한 유저 접근,헤더에 값이 없을때",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+            @ApiResponse(responseCode = "403", description = "권한이 없는 경우",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+            @ApiResponse(responseCode = "404", description = "요청한 URL/URI와 일치하는 항목을 찾지 못함,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+            @ApiResponse(responseCode = "500", description = "외부 API 요청 실패, 정상적 수행을 할 수 없을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
+    })
+    @DeleteMapping("{mediaId}")
+    public void deleteMedia(){
 
     }
-    @GetMapping("{dashboardId}")
-    @Operation(summary = "id별 대시보드 ")
+    @Hidden
+    @Operation(summary = "광고 수정")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "성공"),
             @ApiResponse(responseCode = "400", description = "요청 형식 혹은 요청 콘텐츠가 올바르지 않을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
@@ -42,12 +106,12 @@ public class DashBoardController {
             @ApiResponse(responseCode = "404", description = "요청한 URL/URI와 일치하는 항목을 찾지 못함,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
             @ApiResponse(responseCode = "500", description = "외부 API 요청 실패, 정상적 수행을 할 수 없을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
     })
-    public void findById(){
+    @PatchMapping("{mediaId}")
+    public void updateMedia(){
 
     }
-
-    @GetMapping("media/{mediaId}/board")
-    @Operation(summary = "특정 광고에 대한 대시보드 리스트")
+    @Hidden
+    @Operation(summary = "id 별 광고 조회")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "성공"),
             @ApiResponse(responseCode = "400", description = "요청 형식 혹은 요청 콘텐츠가 올바르지 않을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
@@ -56,20 +120,8 @@ public class DashBoardController {
             @ApiResponse(responseCode = "404", description = "요청한 URL/URI와 일치하는 항목을 찾지 못함,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
             @ApiResponse(responseCode = "500", description = "외부 API 요청 실패, 정상적 수행을 할 수 없을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
     })
+    @GetMapping("{mediaId}")
     public void findByMediaId(){
 
     }
-    @GetMapping("media/{mediaId}/board/{boardId}")
-    @Operation(summary = "특정 광고 신청 단위별 대시보드 ")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "성공"),
-            @ApiResponse(responseCode = "400", description = "요청 형식 혹은 요청 콘텐츠가 올바르지 않을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-            @ApiResponse(responseCode = "401", description = "토큰 시간 만료, 형식 오류,로그아웃한 유저 접근,헤더에 값이 없을때",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-            @ApiResponse(responseCode = "403", description = "권한이 없는 경우",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-            @ApiResponse(responseCode = "404", description = "요청한 URL/URI와 일치하는 항목을 찾지 못함,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-            @ApiResponse(responseCode = "500", description = "외부 API 요청 실패, 정상적 수행을 할 수 없을 때,",content = @Content(schema = @Schema(implementation = ErrorResponse.class))),
-    })
-    public void findByMediaIdAndApplicationId(){}
-
-
 }
