+++ src/server/src/main/java/com/drm/server/controller/dto/request/ApplyRequest.java
@@ -1,7 +1,9 @@
 package com.drm.server.controller.dto.request;
 
 import com.drm.server.domain.mediaApplication.Status;
+import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
+import jakarta.validation.constraints.Pattern;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
@@ -11,14 +13,17 @@ public class ApplyRequest {
     @Getter
     @Setter
     public static class Create{
-
+        @Schema(description = "디스플레이 위치 id")
         private Long locationId;
+        @Schema(description = "시작날짜",example = "2024-04-03")
         private String startDate;
+        @Schema(description = "마감날짜",example = "2024-04-05")
         private String endDate;
     }
     @Getter
     @Setter
     public static class UpdateStatus{
+        @Schema(description = "신청 상태",example = "ACCEPT")
         private Status status;
 
     }
