--- src/server/src/main/java/com/drm/server/controller/dto/request/DashBoardRequest.java
@@ -1,11 +1,0 @@
-package com.drm.server.controller.dto.request;
-
-import jakarta.validation.Valid;
-
-@Valid
-public class DashBoardRequest {
-    public static class Create{
-        private String title;
-        private String discription;
-    }
-}
