--- src/server/src/main/java/com/drm/server/controller/dto/request/MediaApplicationRequest.java
@@ -1,8 +1,0 @@
-package com.drm.server.controller.dto.request;
-
-import jakarta.validation.Valid;
-
-@Valid
-public class MediaApplicationRequest {
-
-}
