+++ src/server/src/main/java/com/drm/server/controller/dto/response/MediaResponse.java
@@ -2,6 +2,7 @@ package com.drm.server.controller.dto.response;
 
 import com.drm.server.domain.dashboard.Dashboard;
 import com.drm.server.domain.media.Media;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 
 @RequiredArgsConstructor
@@ -9,10 +10,13 @@ public class MediaResponse {
     @Getter
     @Setter
     public static class MediaInfo{
+        @Schema(description = "광고 id",example = "1")
         private Long mediaId;
-
+        @Schema(description = "광고 이미지 링크",example = "https://kr.object.ncloudstorage.com/k-eum/image/..")
         private String mediaLink;
+        @Schema(description = "광고 제목",example = "이것은 광고 제목")
         private String title;
+        @Schema(description = "광고 설명",example = "이것은 광고 설명")
         private String description;
         private DashboardResponse.DashboardInfo dashboard;
