+++ src/server/src/main/java/com/drm/server/controller/dto/response/UserResponse.java
@@ -16,6 +16,7 @@ public class UserResponse {
         private Long userId;
         @Schema(description = "이메일",example = "sdkf@gmai.com")
         private String email;
+        @Schema(description = "회사",example = "(주)국민KM")
         private String company;
