+++ src/server/src/main/java/com/drm/server/domain/user/User.java
@@ -33,8 +33,8 @@ public class User extends BaseTimeEntity {
 
     @OneToMany(mappedBy = "user",orphanRemoval = true)
     private List<Dashboard> dashboards = new ArrayList<>();
-    public static User toEntity(String email,String password){
-        return User.builder().email(email).password(password).deleted(false).build();
+    public static User toEntity(String email,String password,String company){
+        return User.builder().email(email).password(password).company(company).deleted(false).build();
     }
 
 }
